refactor: migrate retro functions to retro module
